+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [not yet released]
+
+### Fixed
+- in webGL with brotli-compression for the binary-tiles the accompanying data-files could not be found.
+
 
 ## [0.1.1] - 17-05-2022